Move to shared & tweak.
